+++ operator/pkg/controllers/cluster_controller.go
@@ -1,6 +1,6 @@
 // Package controllers contains k8s controller logic for AIS cluster
 /*
- * Copyright (c) 2021-2024, NVIDIA CORPORATION. All rights reserved.
+ * Copyright (c) 2021-2025, NVIDIA CORPORATION. All rights reserved.
  */
 package controllers
 
@@ -13,7 +13,6 @@ import (
 	"time"
 
 	aiscmn "github.com/NVIDIA/aistore/cmn"
-	aismeta "github.com/NVIDIA/aistore/core/meta"
 	aisv1 "github.com/ais-operator/api/v1beta1"
 	aisclient "github.com/ais-operator/pkg/client"
 	"github.com/ais-operator/pkg/resources/cmn"
@@ -24,7 +23,6 @@ import (
 	"github.com/go-logr/logr"
 	apiv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/equality"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
@@ -986,156 +984,4 @@ func (r *AIStoreReconciler) checkAISClusterReady(ctx context.Context, ais *aisv1
 func (r *AIStoreReconciler) recordError(ctx context.Context, ais *aisv1.AIStore, err error, msg string) {
 	logf.FromContext(ctx).Error(err, msg)
 	r.recorder.Eventf(ais, corev1.EventTypeWarning, EventReasonFailed, "%s, err: %v", msg, err)
-}
-
-func shouldUpdatePodTemplate(desired, current *corev1.PodTemplateSpec) (bool, string) {
-	if len(desired.Spec.Containers) != len(current.Spec.Containers) {
-		return true, "updating desired containers"
-	}
-
-	for _, daemon := range []struct {
-		desiredContainer *corev1.Container
-		currentContainer *corev1.Container
-	}{
-		{&desired.Spec.InitContainers[0], &current.Spec.InitContainers[0]},
-		{&desired.Spec.Containers[0], &current.Spec.Containers[0]},
-	} {
-		if daemon.desiredContainer.Image != daemon.currentContainer.Image {
-			return true, fmt.Sprintf("updating image for %q container", daemon.desiredContainer.Name)
-		}
-		if !equality.Semantic.DeepEqual(daemon.desiredContainer.Env, daemon.currentContainer.Env) {
-			return true, fmt.Sprintf("updating env variables for %q container", daemon.desiredContainer.Name)
-		}
-		if shouldUpdateResources(&daemon.desiredContainer.Resources, &daemon.currentContainer.Resources) {
-			return true, fmt.Sprintf("updating resource requests/limits for %q container", daemon.desiredContainer.Name)
-		}
-	}
-
-	if shouldUpdateAnnotations(desired.Annotations, current.Annotations) {
-		return true, "updating annotations"
-	}
-
-	if !equality.Semantic.DeepEqual(desired.Labels, current.Labels) {
-		return true, "updating labels"
-	}
-
-	// Both `desired.Spec.SecurityContext` and `current.Spec.SecurityContext` are
-	// expected to be non-nil here as `SecurityContext` should be set by default.
-	if !equality.Semantic.DeepEqual(desired.Spec.SecurityContext, current.Spec.SecurityContext) {
-		return true, "updating security context"
-	}
-
-	// We already know desired number of containers matches current here,
-	// so if using sidecar, compare the images of the sidecar container.
-	if len(desired.Spec.Containers) > 1 {
-		if desired.Spec.Containers[1].Image != current.Spec.Containers[1].Image {
-			return true, fmt.Sprintf("updating image for %q container", desired.Spec.Containers[1].Name)
-		}
-	}
-
-	return false, ""
-}
-
-func shouldUpdateResources(desired, current *corev1.ResourceRequirements) bool {
-	// TODO: Remove check in next major version (causes cluster restart)
-	// If we already have ephemeral storage request, do a full comparison
-	if current.Requests.StorageEphemeral() != nil && !current.Requests.StorageEphemeral().IsZero() {
-		return !equality.Semantic.DeepEqual(desired, current)
-	}
-	// Do not sync if the only change is *adding* ephemeral storage request
-	desFiltered := desired.DeepCopy()
-	delete(desFiltered.Requests, corev1.ResourceEphemeralStorage)
-	return !equality.Semantic.DeepEqual(desFiltered, current)
-}
-
-func shouldUpdateAnnotations(desired, current map[string]string) bool {
-	if equality.Semantic.DeepDerivative(desired, current) {
-		return false
-	}
-	restartHash, exists := desired[cmn.RestartConfigHashAnnotation]
-	// At this point annotations are not equal -- If the restart hash does not exist trigger sync
-	if !exists {
-		return true
-	}
-	// If the hash is different and NOT initial, trigger sync
-	nonInitial := !strings.HasSuffix(restartHash, cmn.RestartConfigHashInitial)
-	if nonInitial && restartHash != current[cmn.RestartConfigHashAnnotation] {
-		return true
-	}
-	// Compare the desired to current WITHOUT the restart hash and trigger if not equivalent
-	desiredCopy := make(map[string]string)
-	for k, v := range desired {
-		desiredCopy[k] = v
-	}
-	delete(desiredCopy, cmn.RestartConfigHashAnnotation)
-	return !equality.Semantic.DeepDerivative(desiredCopy, current)
-}
-
-func syncPodTemplate(desired, current *corev1.PodTemplateSpec) (updated bool) {
-	for _, daemon := range []struct {
-		desiredContainer *corev1.Container
-		currentContainer *corev1.Container
-	}{
-		{&desired.Spec.InitContainers[0], &current.Spec.InitContainers[0]},
-		{&desired.Spec.Containers[0], &current.Spec.Containers[0]},
-	} {
-		if equality.Semantic.DeepDerivative(*daemon.desiredContainer, *daemon.currentContainer) {
-			continue
-		}
-		*daemon.currentContainer = *daemon.desiredContainer
-		updated = true
-	}
-
-	if !equality.Semantic.DeepDerivative(desired.Annotations, current.Annotations) {
-		current.Annotations = desired.Annotations
-		updated = true
-	}
-
-	if !equality.Semantic.DeepEqual(desired.Labels, current.Labels) {
-		current.Labels = desired.Labels
-		updated = true
-	}
-
-	if !equality.Semantic.DeepEqual(desired.Spec.SecurityContext, current.Spec.SecurityContext) {
-		current.Spec.SecurityContext = desired.Spec.SecurityContext
-		updated = true
-	}
-
-	if syncSidecarContainer(desired, current) {
-		updated = true
-	}
-
-	return
-}
-
-func findAISNodeByPodName(nodeMap aismeta.NodeMap, podName string) (*aismeta.Snode, error) {
-	for _, node := range nodeMap {
-		if strings.HasPrefix(node.ControlNet.Hostname, podName) {
-			return node, nil
-		}
-	}
-	return nil, fmt.Errorf("no matching AIS node found for pod %q", podName)
-}
-
-func syncSidecarContainer(desired, current *corev1.PodTemplateSpec) (updated bool) {
-	// We have no sidecar, and don't want one
-	if len(desired.Spec.Containers) < 2 && len(current.Spec.Containers) < 2 {
-		return false
-	}
-	// We want to remove the sidecar
-	if len(desired.Spec.Containers) < 2 && len(current.Spec.Containers) > 1 {
-		current.Spec.Containers = current.Spec.Containers[:1]
-		return true
-	}
-	// Add a new sidecar
-	if len(desired.Spec.Containers) > 1 && len(current.Spec.Containers) < 2 {
-		current.Spec.Containers = append(current.Spec.Containers, desired.Spec.Containers[1])
-		return true
-	}
-	// If sidecar is already updated, no change
-	if equality.Semantic.DeepDerivative(desired.Spec.Containers[1], current.Spec.Containers[1]) {
-		return false
-	}
-	current.Spec.Containers[1] = desired.Spec.Containers[1]
-	return true
 }